# services/adapter-fabric/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TANTU Adapter Fabric",
    "description": "Multi-protocol shopfloor ingestion \u2014 OPC-UA, Modbus, MQTT, MTConnect, EtherNet/IP, Camera \u2014 normalized to one schema, derived DefectEvents only (no image field).",
    "version": "0.1.0"
  },
  "paths": {
...
    "/auth/token": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Issue JWT (demo)",
        "operationId": "auth_token_auth_token_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/adapters": {
      "post": {
        "tags": [
          "adapters"
        ],
        "summary": "Register / upsert adapter",
        "operationId": "register_adapter_adapters_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdapterConfigIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "adapters"
        ],
        "summary": "List adapters",
        "operationId": "list_adapters_adapters_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/adapters/{adapter_id}": {
      "get": {
        "tags": [
          "adapters"
        ],
        "summary": "Get Adapter",
        "operationId": "get_adapter_adapters__adapter_id__get",
        "parameters": [
          {
            "name": "adapter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Adapter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "adapters"
        ],
        "summary": "Delete Adapter",
        "operationId": "delete_adapter_adapters__adapter_id__delete",
        "parameters": [
          {
            "name": "adapter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Adapter Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/adapters/{adapter_id}/start": {
      "post": {
        "tags": [
          "adapters"
        ],
        "summary": "Start Adapter",
        "operationId": "start_adapter_adapters__adapter_id__start_post",
        "parameters": [
          {
            "name": "adapter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Adapter Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/adapters/{adapter_id}/stop": {
      "post": {
        "tags": [
          "adapters"
        ],
        "summary": "Stop Adapter",
        "operationId": "stop_adapter_adapters__adapter_id__stop_post",
        "parameters": [
          {
            "name": "adapter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Adapter Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/adapters/{adapter_id}/poll": {
      "post": {
        "tags": [
          "adapters"
        ],
        "summary": "Poll once and return normalized readings",
        "operationId": "poll_adapter_adapters__adapter_id__poll_post",
        "parameters": [
          {
            "name": "adapter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Adapter Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/readings": {
      "get": {
        "tags": [
          "telemetry"
        ],
        "summary": "Poll all adapters once",
        "operationId": "readings_all_readings_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ingest": {
      "post": {
        "tags": [
          "telemetry"
        ],
        "summary": "Direct ingest of a normalized reading (for gateways)",
        "operationId": "ingest_reading_ingest_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestReadingIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/events": {
      "get": {
        "tags": [
          "telemetry"
        ],
        "summary": "Synthetic derived DefectEvents (demo)",
        "operationId": "list_events_events_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 5,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tag-map/preview": {
      "post": {
        "tags": [
          "adapters"
        ],
        "summary": "Preview tag-map compound formula",
        "operationId": "tag_map_preview_tag_map_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompoundPreviewIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AdapterConfigIn": {
        "properties": {
          "adapter_id": {
            "type": "string",
            "pattern": "^[a-z0-9][a-z0-9\\-_]{1,63}$",
            "title": "Adapter Id"
          },
          "protocol": {
            "type": "string",
            "title": "Protocol",
            "description": "opcua | modbus | mqtt | mtconnect | ethernet_ip | camera"
          },
          "station_id": {
            "type": "string",
            "title": "Station Id"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/TagMappingIn"
            },
            "type": "array",
            "title": "Tags"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params"
          },
          "poll_interval_ms": {
            "type": "integer",
            "title": "Poll Interval Ms",
            "default": 1000
          }
        },
        "type": "object",
        "required": [
          "adapter_id",
          "protocol",
          "station_id"
        ],
        "title": "AdapterConfigIn"
      },
      "CompoundPreviewIn": {
        "properties": {
          "formula": {
            "type": "string",
            "title": "Formula"
          },
          "variables": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Variables"
          }
        },
        "type": "object",
        "required": [
          "formula",
          "variables"
        ],
        "title": "CompoundPreviewIn"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IngestReadingIn": {
        "properties": {
          "station_id": {
            "type": "string",
            "title": "Station Id"
          },
          "metric": {
            "type": "string",
            "title": "Metric"
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "unit": {
            "type": "string",
            "title": "Unit",
            "default": ""
          },
          "protocol": {
            "type": "string",
            "title": "Protocol",
            "default": "unknown"
          },
          "adapter_id": {
            "type": "string",
            "title": "Adapter Id",
            "default": ""
          },
          "source_tag": {
            "type": "string",
            "title": "Source Tag",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "station_id",
          "metric",
          "value"
        ],
        "title": "IngestReadingIn"
      },
      "TagMappingIn": {
        "properties": {
          "source_tag": {
            "type": "string",
            "title": "Source Tag"
          },
          "metric": {
            "type": "string",
            "title": "Metric"
          },
          "unit": {
            "type": "string",
            "title": "Unit",
            "default": ""
          },
          "scale": {
            "type": "number",
            "title": "Scale",
            "default": 1.0
          },
          "offset": {
            "type": "number",
            "title": "Offset",
            "default": 0.0
          },
          "data_type": {
            "type": "string",
            "title": "Data Type",
            "default": "float"
          },
          "compound_formula": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compound Formula"
          },
          "source_tags": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Tags"
          }
        },
        "type": "object",
        "required": [
          "source_tag",
          "metric"
        ],
        "title": "TagMappingIn"
      },
      "TokenIn": {
        "properties": {
          "sub": {
            "type": "string",
            "title": "Sub",
            "default": "operator"
          },
          "plant_id": {
            "type": "string",
            "title": "Plant Id",
            "default": "plant-demo-01"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "default": "engineer"
          },
          "exp_min": {
            "type": "integer",
            "title": "Exp Min",
            "default": 60
          }
        },
        "type": "object",
        "title": "TokenIn"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}